Upgrade to fmt8 (#945)
* Upgrade to fmt8

* Start pulling the latest nixos for fmt_8

* Add lock file for latest

* Upgrade to latest Nixos

This keeps the singularity builder on the old version, so we shouldn't
run into issues with the builder

* Add WindowFunction::Form formatter

This will *hopefully* fix the OSX build

* Fix extra colon

Also clean up warnings

* Handle the formatting object at its root

* Cast enumerations to int before passing to fmt

* Fix printing of RegionalDistributor

* Another WindowFunction throw

* Another explicit vector import

* Fix up another format call

* Never import fmt/core.h

That import can interfere with the ability to use the library as
header only.

* Enforce fmt header only

* Add linking back in

I guess I needed this for the headers

* Fix format in configurationviewer

* Fix fmt type error in dataviewer_interaction.cpp

* Fix fmt in siteviewer_interaction

* Future-proof the FMT_HEADER_ONLY tag

I've added it as a compile time definition, instead of needing to
shove it into every source file.  That way, in the future, we can use
fmt in any file without needing to worry about including the define.

* Fix fmt in speciesviewer_interaction.cpp

* Correctly add vector to primitiveinfo.h

+++ src/base/enumoptions.h
@@ -140,7 +140,7 @@ template <class E> class EnumOptions : public EnumOptionsBase
                                [enumeration](auto &option) { return enumeration == option.enumeration(); });
         if (it == options_.cend())
             throw(std::runtime_error(fmt::format(
-                "Enumerated options '{}' missing enumeration {}, so can't return its keyword.\n", name_, enumeration)));
+                "Enumerated options '{}' missing enumeration {}, so can't return its keyword.\n", name_, (int)enumeration)));
 
         return it->keyword();
     }
@@ -162,7 +162,7 @@ template <class E> class EnumOptions : public EnumOptionsBase
         auto it = std::find_if(options_.cbegin(), options_.cend(),
                                [enumeration](auto &option) { return enumeration == option.enumeration(); });
         if (it == options_.cend())
-            throw(std::runtime_error(fmt::format("No option set for enumeration '{}'.\n", enumeration)));
+            throw(std::runtime_error(fmt::format("No option set for enumeration '{}'.\n", (int)enumeration)));
         return *it;
     }
     // Return minimum number of arguments for the specified enumeration
@@ -171,7 +171,7 @@ template <class E> class EnumOptions : public EnumOptionsBase
         auto it = std::find_if(options_.begin(), options_.end(),
                                [enumeration](auto &option) { return enumeration == option.enumeration(); });
         if (it == options_.end())
-            throw(std::runtime_error(fmt::format("No option set for enumeration '{}'.\n", enumeration)));
+            throw(std::runtime_error(fmt::format("No option set for enumeration '{}'.\n", (int)enumeration)));
         return it->minArgs();
     }
     // Return maximum number of arguments for the specified enumeration
@@ -180,7 +180,7 @@ template <class E> class EnumOptions : public EnumOptionsBase
         auto it = std::find_if(options_.begin(), options_.end(),
                                [enumeration](auto &option) { return enumeration == option.enumeration(); });
         if (it == options_.end())
-            throw(std::runtime_error(fmt::format("No option set for enumeration '{}'.\n", enumeration)));
+            throw(std::runtime_error(fmt::format("No option set for enumeration '{}'.\n", (int)enumeration)));
         return it->maxArgs();
     }
     // Check number of arguments provided to keyword
